+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * **BREAKING**: `muuntaja.options` namespace is thrown away. 
   * new helpers in `muuntaja.core`: `transform-formats` & `select-formats`
   * `muuntaja.options/default-options-with-format-regexps` can be copy-pasted from below:
+* default-options support all JVM registered charsets (instead of just `utf-8`)
   
 ```clj
 (def default-options-with-format-regexps